consolidate definitions of print_result_line
