+++ helpers.py
@@ -1,3 +1,5 @@
+from config import SEPARATOR
+
 def sizeof_fmt(num):
     for x in ['bytes','KB','MB','GB','TB']:
         if num < 1024.0:
@@ -17,4 +19,19 @@ def print_header_line():
                 "Num Downloads",
              ]
     print SEPARATOR.join(l)
+
+def print_result_line(c):
+    #c.offer[0].micros/1000000.0
+    #c.offer[0].currencyCode
+    l = [ c.title,
+                c.docid,
+                c.creator,
+                len(c.annotations.badgeForCreator), # Is Super Developer?
+                c.offer[0].formattedAmount,
+                c.offer[0].offerType,
+                c.details.appDetails.versionCode,
+                sizeof_fmt(c.details.appDetails.installationSize),
+                "%.2f" % c.aggregateRating.starRating,
+                c.details.appDetails.numDownloads]
+    print SEPARATOR.join(unicode(i).encode('utf8') for i in l)
